refactor(react-ds-global-form): bindField index.ts barrel-only, impl in bindField.ts
Follow the code standard that an `index.ts` is a barrel (re-exports only): the
implementation moves to `common/bindField/bindField.ts` (default export) and
`index.ts` just re-exports the default + the public types from `./types.js`.
No behaviour change; consumers keep importing the barrel.
